+++ backend/src/routes/kitchen.js
@@ -65,7 +65,7 @@ router.delete('/items/:id', async (req, res) => {
 // FIX: Hardened Withdraw
 router.post('/withdraw', async (req, res) => {
   const { item_id, quantity, reason } = req.body;
-  const qty = parseInt(quantity);
+  const qty = parseFloat(quantity);
   
   if (!item_id || isNaN(qty) || qty <= 0) {
     return res.status(400).json({ error: 'Valid Item ID and quantity are required.' });
@@ -75,7 +75,7 @@ router.post('/withdraw', async (req, res) => {
     const [items] = await pool.query('SELECT quantity FROM kitchen_items WHERE id = ? AND is_active = 1', [item_id]);
     if (!items.length) return res.status(404).json({ error: 'Item not found.' });
     
-    if (parseInt(items[0].quantity) < qty) {
+    if (parseFloat(items[0].quantity) < qty) {
       return res.status(400).json({ error: `Insufficient stock. Available: ${items[0].quantity}` });
     }
 
@@ -96,7 +96,7 @@ router.post('/withdraw', async (req, res) => {
 
 router.post('/restock', async (req, res) => {
   const { item_id, quantity, reason, transaction_date } = req.body;
-  const qty = parseInt(quantity);
+  const qty = parseFloat(quantity);
 
   if (!item_id || isNaN(qty) || qty <= 0) {
     return res.status(400).json({ error: 'Valid Item ID and quantity are required.' });
